(PC-35453)[API] refactor: move start identification tests
Routes defined in `subscription.py` should be in `subscription_test.py`
